templates/key_results/requirements.txt
shimoku==2.1.1
python-dotenv==1.0.1